Fix linting errors
We accept the TypeScript type annotation 'any' for things we pull from
the storage of smart contracts as we have no idea of the structure of
this storage.

Also adds some robustness to the code that fetches the token symbol from
the contract storage. This latter part probably should have been a
separate commit, but since I always fail when using edit option in git
add -p it is part of this commit instead.