Add PDF generation of SK
Add client-side PDF generation of secret key, and attempt to ensure that
the user has saved this PDF file, before the user is allowed to operate
the address. This is done in an attempt to minimize the risk that the
user loses their secret key and thus their funds.

This commit adds a client-side PDF generation library which the client
has to download when visiting this side. As far as I can see, this has
added almost 1MB to the built JS, such that the JS is now 1.4MB. I
consider this size acceptable and the trade-off worth it since it vastly
improves the key-management security of this web wallet.