Add UI to make arbitrary function calls
This adds a modal to make arbitraty function calls on deployed smart
contracts and it includes this modal in the Main component of this
project. This restructures and renames what was previously called the
"Deployment" component as it now contains more than just deployment
functionality.

Tested for the FA1.2 'transfer' endpoint and KISS 'change_activity_log'
endpoints. This functionality was added to make calls to the latter on
already deployed KISS token contracts.