Replace input for number args with value read from blockchain
Instead of letting the user set the number of arguments that a function
on a deployed contract takes, read this value from the blockchain. This
not only reads the number of arguments but also their types. And adds
some rudimentary input validation for these arguments.

We can probably take this input validation further longer, but this is
an appropriate place to stop now.