Remove eslint/TypeScript compile warnings
This is done by upgrading material-table as this upgrades the dependency
jspdf which produced a compile-time warning. Furthermore, some linting
pragmas have been inserted to remove warnings that FA1_2-like names were
illegal for exported components. We like this naming since we cannot
call these components their real name "FA1.2" since dots are not allowed
in names. Also change a '==' to '===' as use of the ill-behaved JS '=='
operator is discouraged.