Add KISS support for FA2 interagration
KISS support is added to the FA2 integration. It reuses functions and
components from the FA1.2 KISS integration that was developed earlier.

In this commit, the FA2 integration is ported from a functional
component to a regular stateful one since I had a problem with multiple
requests to the backend being made. As I am more familiar with these
stateful components, it seemed the easiest way for me to fix that.